+++ src/main/java/com/gaaji/chatmessage/global/config/spring/JsonLocalDateTimeSerializer.java
@@ -1,4 +1,4 @@
-package com.gaaji.chatmessage.global.config;
+package com.gaaji.chatmessage.global.config.spring;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
